Move EnrollButton position
